feat: add Invite button to winger Friends tab
Extracts the invite-a-wingperson flow into a shared
InviteWingpersonSheet component so it can be reused across both modes.

- New InviteWingpersonSheet component (modal, phone form, contacts
  picker, SMS sending) with a variant prop for context-aware copy
- Winger Friends tab gets an Invite button in the header that opens
  the sheet with winger-specific copy: "Help a friend date better"
- Dater Wingpeople screen refactored to use the shared component;
  behaviour unchanged
